\name{ames}
\docType{data}
\alias{ames}
\title{informacoes sobre casas vendidas na cidade de Ames, nos EUA}
\description{
   Este banco de dados apresenta informacoes sobre 2930 casas que foram\dQuote{major}
   vendidas na cidade de Ames, no estado de Iowa, nos Estados Unidos.
}
\usage{ames}
\format{A tibble containing 2930 observations.}
\keyword{datasets}
