% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise.R
\name{analisa}
\alias{analisa}
\title{Funcao de Analise da Base de Dados AMES}
\usage{
analisa()
}
\value{
Um grafico de analise do modelo linear generalizado de previsao do valor de venda dos imoveis em Ames, Iowa, Estados Unidos.
}
\description{
Este banco de dados apresenta informacoes sobre 2930 casas que foram vendidas na cidade de Ames, no estado de Iowa, nos Estados Unidos.
}
