% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conta_NAs.R
\name{pegar_num_NAs}
\alias{pegar_num_NAs}
\title{Funcao de Analise da Base de Dados AMES}
\usage{
pegar_num_NAs(tab, nome_coluna, prop = FALSE)
}
\arguments{
\item{tab}{Tabela ou base de dados que contem a variavel a ser analisada.}

\item{nome_coluna}{Variavel a ser analisada.}

\item{prop}{Opcao de fazer o calculo da proporcao de dados nao disponiveis na variavel.}
}
\value{
Se prop = FALSE, retorna o numero de observacoes com dados nao disponiveis. Caso prop = TRUE, retorna a proporcao de dados nao disponiveis na variavel.
}
\description{
Esta funcao conta o numero de dados nao disponiveis em uma determinada variavel de uma determinada base de dados. Tambem calcula a proporcao de dados nao disponiveis desta variavel.
}
