% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valor_Medio_Venda.R
\name{sumarizar_venda_media}
\alias{sumarizar_venda_media}
\title{Funcao de Analise da Base de Dados AMES}
\usage{
sumarizar_venda_media(tab, nome_coluna)
}
\arguments{
\item{tab}{Tabela ou base de dados que contem a variavel a ser analisada.}

\item{nome_coluna}{Variavel a ser analisada.}
}
\value{
Retorna o valor medio da venda para cada valor da variavel escolhida.
}
\description{
Esta funcao calcula o valor medio de venda por quebras de uma determinada variavel..
}
