\name{ames}
\docType{data}
\alias{ames}
\title{informacoes sobre casas vendidas na cidade de Ames, nos EUA}
\description{
   Este banco de dados apresenta informacoes sobre 2930 casas que foram vendidas na cidade de Ames, no estado de Iowa, nos Estados Unidos.
}
\usage{ames}
\format{Uma tibble contendo 2930 observacoes.}
\keyword{datasets}
